pyproject.toml
[tool.poetry]
name = "simplybot"
version = "0.1.0"
description = "RAG chatbot using LangChain, Qdrant and ElevenLabs"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]

python = "^3.10"
langchain = "^0.1.16"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
langchain-openai = "^0.0.5"
langchain-community = "^0.0.38"
crewai = "^0.11.2"
qdrant-client = "^1.7.0"
streamlit = "^1.28.1"
python-multipart = "^0.0.6"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
openai = "^1.3.7"
elevenlabs = "^0.2.26"
pydub = "^0.25.1"
PyPDF2 = "^3.0.1"
python-docx = "^1.1.0"
requests = "^2.31.0"
sentence-transformers = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mkdocs-material = "^9.6.16"
mkdocs-mermaid2-plugin = "^1.2.1"
mkdocs-awesome-pages-plugin = "^2.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start-api = "uvicorn simplybot.main:app --reload --host 0.0.0.0 --port 8000"
start-gui = "streamlit run simplybot/gui/streamlit_app.py" 